requirements.txt
requests==2.31.0
acryl_datahub==0.13.0
pandas==2.2.1
python-dotenv==1.0.1